feat: add ExecutorManager and ExecutorPool classes for managing executor instances
