+++ src/Classes/ExecutorManager.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Pharaonic\Laravel\Executor\Classes;
+
+class ExecutorManager
+{
+    /**
+     * The executor pool instance.
+     *
+     * @var ExecutorPool
+     */
+    public ExecutorPool $pool;
+
+    public function __construct()
+    {
+        $this->pool = new ExecutorPool();
+    }
+}
